[rename] widgets news 폴더 추가
