+++ src/widgets/news/news-header/ui/NewsHeader.tsx
