+++ src/widgets/news/news-major/index.ts
