+++ src/widgets/news/news-major/ui/MajorNews.tsx
