+++ src/widgets/news/news-section/index.ts
