+++ src/widgets/news/news-section/ui/NewsCategory.tsx
