+++ src/widgets/news/news-section/ui/NewsGrid.tsx
